refactor: типизация пропсов
