+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,10 +1,5 @@
-.tab {
-  position: relative;
-}
-
-.tab:nth-child(1):before {
-  content: "Булки";
-  position: absolute;
+.tabs {
+  display: flex;
 }
 
 .ingredients {
