+++ src/utils/types.ts
@@ -1,0 +1,20 @@
+export interface Ingredient {
+  _id: string;
+  name: string;
+  type: string;
+  proteins: number;
+  fat: number;
+  carbohydrates: number;
+  calories: number;
+  price: number;
+  image: string;
+  image_mobile: string;
+  image_large: string;
+  __v: number;
+}
+
+export interface OpenModalProps {
+  title?: string;
+  inIngredient?: Ingredient | null;
+  inOrder?: boolean;
+}
